The minimum distance is a Double not an Integer
git-svn-id: https://pbtech-vc.med.cornell.edu/public/svn/trunk/qtclustering@14047 60cd2ccc-2b05-0410-9566-da273a940bff

+++ src/edu/cornell/med/icb/clustering/MaxLinkageDistanceCalculator.java
@@ -76,9 +76,9 @@ public abstract class MaxLinkageDistanceCalculator
      * ignoreDistance is returned. The clustering algorithm uses ignoreDistance
      * to recognize cases when the distance is unknown.
      *
-     * @return The minimum integer value, so that max(min, a)=a;
+     * @return The minimum value, so that max(min, a) = a;
      */
     public final double getIgnoreDistance() {
-        return Integer.MIN_VALUE;
+        return Double.MIN_VALUE;
     }
 }
